Move paginator into dedicated modules
Extract the paginator base out of the SDK utils into its own module
(src/paginator.ts, e2b/paginator.py) so it sits next to nothing but its
own concern and owns the ConnectionOpts/ApiParams type dependency directly.
e2b/utils.py is removed (it only held the paginator).
